requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.7
django-elasticsearch-dsl==8.0
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
elastic-transport==8.10.0
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
